Cozmo-10845: V2 Engine and robot process build for android (#4066)
This PR makes it so that `./configure.py build -p android --features standalone` creates a standalone apk using v2 engine and an executable `cozmoRobot2` that runs the robot process. Will eventually need a more kosher deployment strategy, but this is good enough for now. (See standalone-apk/README.md for directions.)

* AndroidHAL.cpp and robot HAL (i.e. halStubs.cpp) largely stubbed out as there is currently no hardware to talk to.
* `GetTimeStamp()` on both engine and robot made to run off `steady_clock` so that images, which are captured engine-side, have timestamps synced with robot. This likely breaks things on robot process but we'll just have to see what happens.
* Prebuild step for robot process called from `configure_engine.py` because ninja didn't like the output-less action called from `cozmoEngine.gyp`
* Need to build `cozmoRobot2` as position-independent executable (PIE) in order to run on device
* Changed `numArgs` from uint8 to int32 in variadic `SendLog` function because of Android complaints
* Tested on Fire HD 8 (5th gen)

+++ robot/espressif/app/application/messages.cpp
@@ -244,7 +244,7 @@ namespace Anki {
     } // namespace Messages
     
     namespace RobotInterface {
-      int SendLog(const LogLevel level, const uint16_t name, const uint16_t formatId, const uint8_t numArgs, ...)
+      int SendLog(const LogLevel level, const uint16_t name, const uint16_t formatId, const int32_t numArgs, ...)
       {
         PrintTrace m;
         if (Messages::missedLogs_ > 0)
